Refactor OpenAI integration and shared GSC utility logic
- Centralize GSC date processing and keyword typing in `app/utils.py`.
- Implement an OpenAI client factory in `app/openai_utils.py`.
- Update all routes to use centralized utilities.
- Re-enable and fix `app/routes/openai.py`.
- Add error handling for AI-driven features.
- Clean up `app/__init__.py` by removing stale commented-out code.
