+++ app/routes/__init__.py
@@ -1,5 +1,5 @@
 from app.routes.default_routes import *
-#from app.routes.openai import *
+from app.routes.openai import *
 from app.routes.plotly import *
 from app.routes.datatables import *
 from app.routes.customer_support import *
